Move constants from EpochEndOffset to OffsetsForLeaderEpochResponse
